controller: Split code into manager.go
